+++ plaster/exceptions.py
@@ -1,4 +1,10 @@
 class NoSectionError(ValueError):
+    """
+    Raised by a :class:`plaster.Loader` which cannot find a section.
+
+    :ivar section: The name of the section that does not exist.
+
+    """
     def __init__(self, section, message=None):
         if message is None:
             message = 'Could not find requested section "{0}".'.format(section)
@@ -8,6 +14,12 @@ class NoSectionError(ValueError):
 
 
 class InvalidURI(ValueError):
+    """
+    Raised by :func:`plaster.parse_uri` when failing to parse a ``config_uri``.
+
+    :ivar uri: The user-supplied ``config_uri`` string.
+
+    """
     def __init__(self, uri, message=None):
         if message is None:
             message = 'Unable to parse config_uri "{0}".'.format(uri)
@@ -17,6 +29,13 @@ class InvalidURI(ValueError):
 
 
 class LoaderNotFound(ValueError):
+    """
+    Raised by :func:`plaster.get_loader` when no loaders match the requested
+    ``scheme``.
+
+    :ivar scheme: The scheme being matched.
+
+    """
     def __init__(self, scheme, message=None):
         if message is None:
             message = (
@@ -28,6 +47,14 @@ class LoaderNotFound(ValueError):
 
 
 class MultipleLoadersFound(ValueError):
+    """
+    Raised by :func:`plaster.get_loader` when more than one loader matches the
+    requested ``scheme``.
+
+    :ivar scheme: The scheme being matched.
+    :ivar loaders: A list of loader entry points.
+
+    """
     def __init__(self, scheme, loaders, message=None):
         if message is None:
             message = (