+++ CHANGES.rst
@@ -4,7 +4,7 @@ unreleased
 - Allow ``config_uri`` syntax ``scheme:path`` alongside ``scheme://path``.
   See https://github.com/mmerickel/plaster/issues/3
 
-- Improve error messages to show the user-supplied values in the error output.
+- Improve errors to show the user-supplied values in the error message.
   See https://github.com/mmerickel/plaster/pull/4
 
 0.1.0 (2016-06-12)